feat: Improved Resource
BREAKING CHANGE:
- Removed EntityRecord
- Removed SimpleResource
- New BaseResource
- RestFetch, FetchMutate, FetchGet first arg is entire args instead of
just params

fix: imports
