+++ packages/rest/src/types.ts
@@ -1,20 +1,23 @@
-import { EndpointInstance, FetchFunction, Schema } from '@rest-hooks/endpoint';
+import type {
+  EndpointInstance,
+  FetchFunction,
+  Schema,
+} from '@rest-hooks/endpoint';
 
-export type RestFetch<
-  P = any,
-  B = RequestInit['body'] | Record<string, any>,
-  R = any,
-> = (this: RestEndpoint, params?: P, body?: B, ...rest: any) => Promise<R>;
-
-export type FetchMutate<
-  P = any,
-  B = RequestInit['body'] | Record<string, any>,
-  R = any,
-> = (this: RestEndpoint, params: P, body: B) => Promise<R>;
-
-export type FetchGet<P = any, R = any> = (
+export type RestFetch<A extends readonly any[] = any[], R = any> = (
   this: RestEndpoint,
-  params: P,
+  ...args: A
+) => Promise<R>;
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type FetchMutate<A extends readonly any[] = [any, {}], R = any> = (
+  this: RestEndpoint,
+  ...args: A
+) => Promise<R>;
+
+export type FetchGet<A extends readonly any[] = [any], R = any> = (
+  this: RestEndpoint,
+  ...args: A
 ) => Promise<R>;
 
 /** Endpoint from a Resource
